# Akka.Remote Performance - String.Format logging perf fix