Fixes #1539 

The log message is being formatted in the default pipeline, even if logging is not enabled. 
This PR moves the message formatting into the conditionals that logs.