@Aaronontheweb 2015-12-14
@rogeralsing could you rebase this on dev and re-run it so the Akka.Remote benchmarks show up?