@Aaronontheweb 2015-12-14
Perf from this build thus far

# Akka.Remote.Tests.Performance.TestTransportRemoteMessagingThroughputSpec+OneWay
__Measures the throughput of Akka.Remote over a particular transport using one-way messaging__
_12/14/2015 9:11:16 PM_
### System Info
```ini
NBench=NBench, Version=0.1.5.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=4
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=4
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=13, MaximumRunTime=00:00:01
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            7.00 |            6.85 |            6.00 |            0.38 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        2,467.00 |        2,371.00 |        2,147.00 |           93.29 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            7.02 |            6.84 |            5.99 |            0.38 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        2,463.66 |        2,368.21 |        2,143.45 |           93.87 |

### Raw Data
#### TotalCollections [Gen0]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            7.00 |            6.99 |  143,032,242.86 |
               2 |            7.00 |            6.98 |  143,171,014.29 |
               3 |            7.00 |            6.99 |  143,154,700.00 |
               4 |            7.00 |            7.02 |  142,376,628.57 |
               5 |            7.00 |            6.99 |  143,014,000.00 |
               6 |            7.00 |            6.99 |  143,138,828.57 |
               7 |            7.00 |            6.99 |  143,011,185.71 |
               8 |            7.00 |            6.99 |  143,108,900.00 |
               9 |            7.00 |            6.99 |  143,050,757.14 |
              10 |            6.00 |            5.99 |  166,881,866.67 |
              11 |            7.00 |            6.99 |  143,022,057.14 |
              12 |            7.00 |            6.99 |  143,138,942.86 |
              13 |            6.00 |            5.99 |  166,943,016.67 |

#### TotalCollections [Gen1]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,225,700.00 |
               2 |            0.00 |            0.00 |1,002,197,100.00 |
               3 |            0.00 |            0.00 |1,002,082,900.00 |
               4 |            0.00 |            0.00 |  996,636,400.00 |
               5 |            0.00 |            0.00 |1,001,098,000.00 |
               6 |            0.00 |            0.00 |1,001,971,800.00 |
               7 |            0.00 |            0.00 |1,001,078,300.00 |
               8 |            0.00 |            0.00 |1,001,762,300.00 |
               9 |            0.00 |            0.00 |1,001,355,300.00 |
              10 |            0.00 |            0.00 |1,001,291,200.00 |
              11 |            0.00 |            0.00 |1,001,154,400.00 |
              12 |            0.00 |            0.00 |1,001,972,600.00 |
              13 |            0.00 |            0.00 |1,001,658,100.00 |

#### TotalCollections [Gen2]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,225,700.00 |
               2 |            0.00 |            0.00 |1,002,197,100.00 |
               3 |            0.00 |            0.00 |1,002,082,900.00 |
               4 |            0.00 |            0.00 |  996,636,400.00 |
               5 |            0.00 |            0.00 |1,001,098,000.00 |
               6 |            0.00 |            0.00 |1,001,971,800.00 |
               7 |            0.00 |            0.00 |1,001,078,300.00 |
               8 |            0.00 |            0.00 |1,001,762,300.00 |
               9 |            0.00 |            0.00 |1,001,355,300.00 |
              10 |            0.00 |            0.00 |1,001,291,200.00 |
              11 |            0.00 |            0.00 |1,001,154,400.00 |
              12 |            0.00 |            0.00 |1,001,972,600.00 |
              13 |            0.00 |            0.00 |1,001,658,100.00 |

#### [Counter] RemoteMessageReceived
           Run # |      operations |  operations / s | ns / operations |
---------------- |---------------- |---------------- |---------------- |
               1 |        2,466.00 |        2,462.98 |      406,012.04 |
               2 |        2,351.00 |        2,345.85 |      426,285.45 |
               3 |        2,335.00 |        2,330.15 |      429,157.56 |
               4 |        2,409.00 |        2,417.13 |      413,713.74 |
               5 |        2,417.00 |        2,414.35 |      414,190.32 |
               6 |        2,331.00 |        2,326.41 |      429,846.33 |
               7 |        2,466.00 |        2,463.34 |      405,952.27 |
               8 |        2,362.00 |        2,357.84 |      424,116.13 |
               9 |        2,467.00 |        2,463.66 |      405,900.00 |
              10 |        2,255.00 |        2,252.09 |      444,031.57 |
              11 |        2,365.00 |        2,362.27 |      423,321.10 |
              12 |        2,452.00 |        2,447.17 |      408,634.83 |
              13 |        2,147.00 |        2,143.45 |      466,538.47 |