@Aaronontheweb 2015-12-15
So I tried running this locally, to see if Azure's CPU throttling was to blame for some of the skew and disparity between the two tests, and this time got a convergent result:

# With Change
----

# Akka.Remote.Tests.Performance.TestTransportRemoteMessagingThroughputSpec+OneWay
__Measures the throughput of Akka.Remote over a particular transport using one-way messaging__
_12/15/2015 5:57:26 PM_
### System Info
```ini
NBench=NBench, Version=0.1.3.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=8
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=8
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=13, MaximumRunTime=00:00:01
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        3,865.00 |        3,833.85 |        3,606.00 |           69.35 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        3,860.73 |        3,829.71 |        3,599.78 |           69.88 |

### Raw Data
#### TotalCollections [Gen0]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,136,226.60 |
               2 |            0.00 |            0.00 |1,001,546,328.71 |
               3 |            0.00 |            0.00 |1,001,727,646.01 |
               4 |            0.00 |            0.00 |1,000,869,809.90 |
               5 |            0.00 |            0.00 |1,000,723,558.68 |
               6 |            0.00 |            0.00 |1,001,072,081.33 |
               7 |            0.00 |            0.00 |1,000,947,639.50 |
               8 |            0.00 |            0.00 |1,001,230,306.33 |
               9 |            0.00 |            0.00 |1,000,856,980.84 |
              10 |            0.00 |            0.00 |1,000,842,868.88 |
              11 |            0.00 |            0.00 |1,000,787,703.95 |
              12 |            0.00 |            0.00 |1,001,217,049.64 |
              13 |            0.00 |            0.00 |1,001,106,292.14 |

#### TotalCollections [Gen1]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,136,226.60 |
               2 |            0.00 |            0.00 |1,001,546,328.71 |
               3 |            0.00 |            0.00 |1,001,727,646.01 |
               4 |            0.00 |            0.00 |1,000,869,809.90 |
               5 |            0.00 |            0.00 |1,000,723,558.68 |
               6 |            0.00 |            0.00 |1,001,072,081.33 |
               7 |            0.00 |            0.00 |1,000,947,639.50 |
               8 |            0.00 |            0.00 |1,001,230,306.33 |
               9 |            0.00 |            0.00 |1,000,856,980.84 |
              10 |            0.00 |            0.00 |1,000,842,868.88 |
              11 |            0.00 |            0.00 |1,000,787,703.95 |
              12 |            0.00 |            0.00 |1,001,217,049.64 |
              13 |            0.00 |            0.00 |1,001,106,292.14 |

#### TotalCollections [Gen2]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,136,226.60 |
               2 |            0.00 |            0.00 |1,001,546,328.71 |
               3 |            0.00 |            0.00 |1,001,727,646.01 |
               4 |            0.00 |            0.00 |1,000,869,809.90 |
               5 |            0.00 |            0.00 |1,000,723,558.68 |
               6 |            0.00 |            0.00 |1,001,072,081.33 |
               7 |            0.00 |            0.00 |1,000,947,639.50 |
               8 |            0.00 |            0.00 |1,001,230,306.33 |
               9 |            0.00 |            0.00 |1,000,856,980.84 |
              10 |            0.00 |            0.00 |1,000,842,868.88 |
              11 |            0.00 |            0.00 |1,000,787,703.95 |
              12 |            0.00 |            0.00 |1,001,217,049.64 |
              13 |            0.00 |            0.00 |1,001,106,292.14 |

#### [Counter] RemoteMessageReceived
           Run # |      operations |  operations / s | ns / operations |
---------------- |---------------- |---------------- |---------------- |
               1 |        3,864.00 |        3,859.61 |      259,093.23 |
               2 |        3,864.00 |        3,858.03 |      259,199.36 |
               3 |        3,606.00 |        3,599.78 |      277,794.69 |
               4 |        3,838.00 |        3,834.66 |      260,779.00 |
               5 |        3,847.00 |        3,844.22 |      260,130.90 |
               6 |        3,851.00 |        3,846.88 |      259,951.20 |
               7 |        3,835.00 |        3,831.37 |      261,003.30 |
               8 |        3,861.00 |        3,856.26 |      259,318.91 |
               9 |        3,864.00 |        3,860.69 |      259,020.96 |
              10 |        3,844.00 |        3,840.76 |      260,364.95 |
              11 |        3,840.00 |        3,836.98 |      260,621.80 |
              12 |        3,861.00 |        3,856.31 |      259,315.48 |
              13 |        3,865.00 |        3,860.73 |      259,018.45 |

# Without change
---
# Akka.Remote.Tests.Performance.TestTransportRemoteMessagingThroughputSpec+OneWay
__Measures the throughput of Akka.Remote over a particular transport using one-way messaging__
_12/15/2015 6:11:02 PM_
### System Info
```ini
NBench=NBench, Version=0.1.3.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=8
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=8
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=13, MaximumRunTime=00:00:01
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            1.00 |            1.00 |            1.00 |            0.00 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        3,776.00 |        3,753.15 |        3,654.00 |           32.41 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            1.00 |            1.00 |            0.99 |            0.00 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        3,771.49 |        3,747.69 |        3,635.49 |           36.09 |

### Raw Data
#### TotalCollections [Gen0]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            1.00 |            1.00 |1,001,084,910.38 |
               2 |            1.00 |            0.99 |1,005,092,279.39 |
               3 |            1.00 |            1.00 |1,000,759,052.39 |
               4 |            1.00 |            1.00 |1,000,974,580.51 |
               5 |            1.00 |            1.00 |1,001,275,208.02 |
               6 |            1.00 |            1.00 |1,001,190,963.90 |
               7 |            1.00 |            1.00 |1,001,339,780.93 |
               8 |            1.00 |            1.00 |1,000,931,389.36 |
               9 |            1.00 |            1.00 |1,001,134,516.06 |
              10 |            1.00 |            1.00 |1,001,099,877.61 |
              11 |            1.00 |            1.00 |1,001,215,339.10 |
              12 |            1.00 |            1.00 |1,001,450,110.80 |
              13 |            1.00 |            1.00 |1,001,535,210.20 |

#### TotalCollections [Gen1]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,084,910.38 |
               2 |            0.00 |            0.00 |1,005,092,279.39 |
               3 |            0.00 |            0.00 |1,000,759,052.39 |
               4 |            0.00 |            0.00 |1,000,974,580.51 |
               5 |            0.00 |            0.00 |1,001,275,208.02 |
               6 |            0.00 |            0.00 |1,001,190,963.90 |
               7 |            0.00 |            0.00 |1,001,339,780.93 |
               8 |            0.00 |            0.00 |1,000,931,389.36 |
               9 |            0.00 |            0.00 |1,001,134,516.06 |
              10 |            0.00 |            0.00 |1,001,099,877.61 |
              11 |            0.00 |            0.00 |1,001,215,339.10 |
              12 |            0.00 |            0.00 |1,001,450,110.80 |
              13 |            0.00 |            0.00 |1,001,535,210.20 |

#### TotalCollections [Gen2]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,084,910.38 |
               2 |            0.00 |            0.00 |1,005,092,279.39 |
               3 |            0.00 |            0.00 |1,000,759,052.39 |
               4 |            0.00 |            0.00 |1,000,974,580.51 |
               5 |            0.00 |            0.00 |1,001,275,208.02 |
               6 |            0.00 |            0.00 |1,001,190,963.90 |
               7 |            0.00 |            0.00 |1,001,339,780.93 |
               8 |            0.00 |            0.00 |1,000,931,389.36 |
               9 |            0.00 |            0.00 |1,001,134,516.06 |
              10 |            0.00 |            0.00 |1,001,099,877.61 |
              11 |            0.00 |            0.00 |1,001,215,339.10 |
              12 |            0.00 |            0.00 |1,001,450,110.80 |
              13 |            0.00 |            0.00 |1,001,535,210.20 |

#### [Counter] RemoteMessageReceived
           Run # |      operations |  operations / s | ns / operations |
---------------- |---------------- |---------------- |---------------- |
               1 |        3,750.00 |        3,745.94 |      266,955.98 |
               2 |        3,654.00 |        3,635.49 |      275,066.31 |
               3 |        3,766.00 |        3,763.14 |      265,735.28 |
               4 |        3,757.00 |        3,753.34 |      266,429.22 |
               5 |        3,730.00 |        3,725.25 |      268,438.39 |
               6 |        3,775.00 |        3,770.51 |      265,216.15 |
               7 |        3,776.00 |        3,770.95 |      265,185.32 |
               8 |        3,775.00 |        3,771.49 |      265,147.39 |
               9 |        3,764.00 |        3,759.73 |      265,976.23 |
              10 |        3,758.00 |        3,753.87 |      266,391.67 |
              11 |        3,765.00 |        3,760.43 |      265,927.05 |
              12 |        3,770.00 |        3,764.54 |      265,636.63 |
              13 |        3,751.00 |        3,745.25 |      267,004.85 |